Make parametric controlled operators inherit from `ControlledOp`  (#5069)
**Context:**

Many controlled operations have been added to PennyLane over the years,
such as CNOT, CRY, CZ, etc. For code tidiness, we would like to make
them all inherit from the
[ControlledOp](https://docs.pennylane.ai/en/stable/code/api/pennylane.ops.op_math.ControlledOp.html)
base.

**Description of the Change:**

`CRX`, `CRY`, `CRZ`, `CRot`, `ControlledPhaseShift` are now:

1. Moved from `pennylane.ops.qubit` to `pennylane.ops.op_math`
2. Inherit from `ControlledOp`

Calling `qml.ctrl` on `RX`, `RY`, `RZ`, `Rot`, and `PhaseShift` with a
single control wire will return gates of types `CRX`, `CRY`, etc. as
opposed to a general `Controlled` operator.

The decomposition of an operator created with calling `qml.ctrl` on
`RX`, `RY`, `RZ`, `Rot`, `PhaseShift` with a single control wire will
now be the full decomposition instead of a single controlled gate. For
example:
  ```
  >>> qml.ctrl(qml.RX(0.123, wires=1), control=0).decomposition()
  [CRX(0.123, wires=[0, 1])]
  ```
  Will become:
  ```
  >>> qml.ctrl(qml.RX(0.123, wires=1), control=0).decomposition()
  [
    RZ(1.5707963267948966, wires=[1]), 
    RY(0.0615, wires=[1]), 
    CNOT(wires=[0, 1]), 
    RY(-0.0615, wires=[1]), 
    CNOT(wires=[0, 1]), 
    RZ(-1.5707963267948966, wires=[1])
  ]
  ```

**Benefits:**

Code cleanliness.

**Related GitHub Issues:**

https://github.com/PennyLaneAI/pennylane/issues/1447

---------
