+++ pennylane/ops/op_math/__init__.py
@@ -64,6 +64,11 @@ Controlled Operator Classes
     ~ControlledQubitUnitary
     ~CY
     ~CZ
+    ~CRX
+    ~CRY
+    ~CRZ
+    ~CRot
+    ~ControlledPhaseShift
 
 Decompositions
 ~~~~~~~~~~~~~~
@@ -94,7 +99,17 @@ from .adjoint import Adjoint, adjoint
 from .composite import CompositeOp
 from .condition import cond, Conditional
 from .controlled import Controlled, ControlledOp, ctrl
-from .controlled_ops import ControlledQubitUnitary, CY, CZ
+from .controlled_ops import (
+    ControlledQubitUnitary,
+    ControlledPhaseShift,
+    CPhase,
+    CRot,
+    CRX,
+    CRY,
+    CRZ,
+    CY,
+    CZ,
+)
 from .decompositions import one_qubit_decomposition, two_qubit_decomposition, sk_decomposition
 from .evolution import Evolution
 from .exp import Exp, exp
@@ -105,4 +120,14 @@ from .sum import Sum, sum
 from .symbolicop import ScalarSymbolicOp, SymbolicOp
 from .controlled_decompositions import ctrl_decomp_zyz, ctrl_decomp_bisect
 
-controlled_qubit_ops = {"ControlledQubitUnitary", "CY", "CZ"}
+controlled_qubit_ops = {
+    "ControlledQubitUnitary",
+    "CY",
+    "CZ",
+    "CRX",
+    "CRY",
+    "CRZ",
+    "CRot",
+    "ControlledPhaseShift",
+    "CPhase",
+}
