+++ tests/ops/op_math/test_controlled_ops.py
@@ -12,36 +12,40 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """
-Unit tests for Operators inheriting from Controlled.
+Unit tests for Operators inheriting from ControlledOp.
 """
 
-import copy
+import functools
 
 import numpy as np
 import pytest
 from scipy.linalg import fractional_matrix_power
-from scipy.sparse import csr_matrix
 from scipy.stats import unitary_group
 
-from gate_data import (
-    CY,
-    CZ,
-)
+from gate_data import CY, CZ, CRotx, CRoty, CRotz, CRot3, ControlledPhaseShift
 
 import pennylane as qml
 from pennylane.wires import Wires
-from pennylane.operation import AnyWires
 from pennylane.ops.qubit.matrix_ops import QubitUnitary
 
-# Non-parametrized operations and their matrix representation
 NON_PARAMETRIZED_OPERATIONS = [
     (qml.CY, CY),
     (qml.CZ, CZ),
 ]
 
-SPARSE_MATRIX_SUPPORTED_OPERATIONS = (
-    (qml.CY(wires=[0, 1]), CY),
-    (qml.CZ(wires=[0, 1]), CZ),
+PARAMETRIZED_OPERATIONS = [
+    (qml.CRX, CRotx),
+    (qml.CRY, CRoty),
+    (qml.CRZ, CRotz),
+    (qml.CRot, CRot3),
+    (qml.ControlledPhaseShift, ControlledPhaseShift),
+]
+
+ALL_OPERATIONS = NON_PARAMETRIZED_OPERATIONS + PARAMETRIZED_OPERATIONS
+
+NON_PARAMETRIC_OPS_DECOMPOSITIONS = (
+    (qml.CY, [qml.CRY(np.pi, wires=[0, 1]), qml.S(0)]),
+    (qml.CZ, [qml.ControlledPhaseShift(np.pi, wires=[0, 1])]),
 )
 
 X = np.array([[0, 1], [1, 0]])
@@ -453,176 +457,495 @@ class TestControlledQubitUnitary:
             )
 
 
-class TestOperations:
-    @pytest.mark.parametrize("op_cls, _", NON_PARAMETRIZED_OPERATIONS)
-    def test_nonparametrized_op_copy(self, op_cls, _, tol):
-        """Tests that copied nonparametrized ops function as expected"""
-        op = op_cls(wires=0 if op_cls.num_wires is AnyWires else range(op_cls.num_wires))
-        copied_op = copy.copy(op)
-        assert qml.equal(copied_op, op, atol=tol, rtol=0)
-        assert copied_op is not op
-        assert qml.equal(copied_op, op, atol=tol, rtol=0)
-
-    @pytest.mark.parametrize("ops, mat", NON_PARAMETRIZED_OPERATIONS)
-    def test_matrices(self, ops, mat, tol):
-        """Test matrices of non-parametrized operations are correct"""
-        op = ops(wires=0 if ops.num_wires is AnyWires else range(ops.num_wires))
-        res_static = op.compute_matrix()
-        res_dynamic = op.matrix()
-        assert np.allclose(res_static, mat, atol=tol, rtol=0)
-        assert np.allclose(res_dynamic, mat, atol=tol, rtol=0)
-
-
-class TestDecompositions:  # pylint: disable=too-few-public-methods
-    def test_CY_decomposition(self, tol):
-        """Tests that the decomposition of the CY gate is correct"""
-        op = qml.CY(wires=[0, 1])
-        gate1, gate2 = op.decomposition()
-        decomposed_matrix = qml.matrix(op.decomposition)()
-
-        assert qml.equal(gate1, qml.CRY(np.pi, wires=[0, 1]))
-        assert qml.equal(gate2, qml.S(0))
-        assert np.allclose(decomposed_matrix, op.matrix(), atol=tol, rtol=0)
-
-    def test_CZ_decomposition(self, tol):
-        """Tests that the decomposition of the CZ gate is correct"""
-        op = qml.CZ(wires=[0, 1])
-        res = op.decomposition()
-
-        assert op.has_decomposition
-        assert len(res) == 1
-        assert qml.equal(qml.ctrl(qml.PhaseShift(np.pi, wires=1), 0), res[0], atol=tol, rtol=0)
-
-        decomposed_matrix = res[0].matrix()
-        assert np.allclose(decomposed_matrix, op.matrix(), atol=tol, rtol=0)
-
-
-class TestEigenval:  # pylint: disable=too-few-public-methods
-    def test_CZ_eigenval(self):
-        """Tests that the CZ eigenvalue matches the numpy eigenvalues of the CZ matrix"""
-        op = qml.CZ(wires=[0, 1])
-        exp = np.linalg.eigvals(op.matrix())
-        res = op.eigvals()
-        assert np.allclose(res, exp)
-
-
-period_two_ops = (
-    qml.CY(wires=(0, 1)),
-    qml.CZ(wires=(0, 1)),
-)
-
-
-class TestPowMethod:
-    @pytest.mark.parametrize("op", period_two_ops)
-    @pytest.mark.parametrize("n", (1, 5, -1, -5))
-    def test_period_two_pow_odd(self, op, n):
-        """Test that ops with a period of 2 raised to an odd power are the same as the original op."""
-        assert qml.equal(op.pow(n)[0], op)
-        assert np.allclose(op.pow(n)[0].matrix(), op.matrix())
-        assert op.pow(n)[0].name == op.name
-
-    @pytest.mark.parametrize("op", period_two_ops)
-    @pytest.mark.parametrize("n", (2, 6, 0, -2))
-    def test_period_two_pow_even(self, op, n):
-        """Test that ops with a period of 2 raised to an even power are empty lists."""
-        assert len(op.pow(n)) == 0
-
-    @pytest.mark.parametrize("op", period_two_ops)
-    def test_period_two_noninteger_power(self, op):
-        """Test that ops with a period of 2 raised to a non-integer power raise an error."""
-        if isinstance(op, (qml.PauliZ, qml.CZ)):
-            pytest.skip("PauliZ can be raised to any power.")
-        with pytest.raises(qml.operation.PowUndefinedError):
-            op.pow(1.234)
-
-        if op.__class__ is qml.CZ:
-            pytest.skip("CZ can be raised to any power.")
-        with pytest.raises(qml.operation.PowUndefinedError):
-            op.pow(1.234)
-
-    @pytest.mark.parametrize("n", (0.12, -3.462, 3.693))
-    def test_cz_general_power(self, n):
-        """Check that CZ raised to an non-integer power that's not the square root
-        results in a controlled PhaseShift."""
-        op_pow = qml.CZ(wires=[0, 1]).pow(n)
-
-        assert len(op_pow) == 1
-        assert isinstance(op_pow[0], qml.ops.ControlledOp)
-        assert isinstance(op_pow[0].base, qml.PhaseShift)
-        assert qml.math.allclose(op_pow[0].data[0], np.pi * (n % 2))
-
-
-class TestControlledMethod:  # pylint: disable=too-few-public-methods
-    """Tests for the _controlled method of non-parametric operations."""
-
-    def test_CZ(self):
-        """Test the PauliZ _controlled method."""
-        out = qml.CZ(wires=[0, 1])._controlled("a")  # pylint: disable=protected-access
-        assert qml.equal(out, qml.CCZ(("a", 0, 1)))
-
-
-class TestSparseMatrix:  # pylint: disable=too-few-public-methods
-    @pytest.mark.parametrize("op, mat", SPARSE_MATRIX_SUPPORTED_OPERATIONS)
-    def test_sparse_matrix(self, op, mat):
-        """Tests the sparse matrix method for operations which support it."""
-        expected_sparse_mat = csr_matrix(mat)
-        sparse_mat = op.sparse_matrix()
-
-        assert isinstance(sparse_mat, csr_matrix)
-        assert isinstance(expected_sparse_mat, csr_matrix)
-        assert all(sparse_mat.data == expected_sparse_mat.data)
-        assert all(sparse_mat.indices == expected_sparse_mat.indices)
-
-
-label_data = [
-    (qml.CY(wires=(0, 1)), "Y"),
-    (qml.CZ(wires=(0, 1)), "Z"),
-]
-
-
-@pytest.mark.parametrize("op, label", label_data)
-def test_label_method(op, label):
-    """Tests that the label method gives the expected result."""
-    assert op.label() == label
-    assert op.label(decimals=2) == label
-
-
-control_data = [
-    (qml.CY(wires=(0, 1)), Wires(0)),
-    (qml.CZ(wires=(0, 1)), Wires(0)),
-]
-
-
-@pytest.mark.parametrize("op, control_wires", control_data)
-def test_control_wires(op, control_wires):
-    """Test ``control_wires`` attribute for non-parametrized operations."""
-
-    assert op.control_wires == control_wires
-
-
-involution_ops = [  # ops who are their own inverses
-    qml.CY((0, 1)),
-    qml.CZ(wires=(0, 1)),
-]
-
-
-@pytest.mark.parametrize("op", involution_ops)
-def test_adjoint_method(op):
-    """Tests the adjoint method for operations that are their own adjoint."""
-    adj_op = copy.copy(op)
-    for _ in range(4):
-        adj_op = adj_op.adjoint()
-
-        assert qml.equal(adj_op, op)
-
-
 @pytest.mark.parametrize("op_cls, _", NON_PARAMETRIZED_OPERATIONS)
-def test_map_wires(op_cls, _):
-    """Test that we can get and set private wires in all operations."""
+def test_map_wires_non_parametric(op_cls, _):
+    """Test get and set private wires in non-parametric controlled operations."""
 
     op = op_cls(wires=[0, 1])
     assert op.wires == Wires((0, 1))
 
     op = op.map_wires(wire_map={0: "a", 1: "b"})
-    assert op.base.wires == Wires(("b"))
-    assert op.control_wires == Wires(("a"))
+    assert op.base.wires == Wires("b")
+    assert op.control_wires == Wires("a")
+
+
+def test_controlled_phase_shift_alias():
+    """Tests that the alias for ControlledPhaseShift works"""
+    assert qml.equal(qml.ControlledPhaseShift(0.123, wires=[0, 1]), qml.CPhase(0.123, wires=[0, 1]))
+
+
+def _arbitrary_crot(x, y, z):
+    """controlled arbitrary single qubit rotation"""
+    c = np.cos(y / 2)
+    s = np.sin(y / 2)
+    return np.array(
+        [
+            [1, 0, 0, 0],
+            [0, 1, 0, 0],
+            [0, 0, np.exp(-0.5j * (x + z)) * c, -np.exp(0.5j * (x - z)) * s],
+            [0, 0, np.exp(-0.5j * (x - z)) * s, np.exp(0.5j * (x + z)) * c],
+        ]
+    )
+
+
+def _arbitrary_crot_broadcasted(x, y, z):
+    """controlled arbitrary single qubit rotation broadcasted"""
+
+    c = np.cos(y / 2)
+    s = np.sin(y / 2)
+    return np.array(
+        [
+            [
+                [1, 0, 0, 0],
+                [0, 1, 0, 0],
+                [0, 0, np.exp(-0.5j * (_x + _z)) * _c, -np.exp(0.5j * (_x - _z)) * _s],
+                [0, 0, np.exp(-0.5j * (_x - _z)) * _s, np.exp(0.5j * (_x + _z)) * _c],
+            ]
+            for _x, _z, _c, _s in zip(x, z, c, s)
+        ]
+    )
+
+
+def _phase_shift_matrix_broadcasted(phi):
+    """Phase shift matrix broadcasted"""
+
+    expected = np.array([np.eye(4, dtype=complex)] * 3)
+    expected[..., 3, 3] = np.exp(1j * phi)
+    return expected
+
+
+EXPECTED_MATRICES = [
+    (qml.CRX, [0], np.identity(4)),
+    (
+        qml.CRX,
+        [np.pi / 2],
+        np.array(
+            [
+                [1, 0, 0, 0],
+                [0, 1, 0, 0],
+                [0, 0, 1 / np.sqrt(2), -1j / np.sqrt(2)],
+                [0, 0, -1j / np.sqrt(2), 1 / np.sqrt(2)],
+            ]
+        ),
+    ),
+    (qml.CRX, [np.pi], np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, -1j], [0, 0, -1j, 0]])),
+    (
+        qml.CRX,
+        [np.array([np.pi / 2, np.pi])],
+        np.array(
+            [
+                [
+                    [1, 0, 0, 0],
+                    [0, 1, 0, 0],
+                    [0, 0, 1 / np.sqrt(2), -1j / np.sqrt(2)],
+                    [0, 0, -1j / np.sqrt(2), 1 / np.sqrt(2)],
+                ],
+                [[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, -1j], [0, 0, -1j, 0]],
+            ]
+        ),
+    ),
+    (qml.CRY, [0], np.identity(4)),
+    (
+        qml.CRY,
+        [np.pi / 2],
+        np.array(
+            [
+                [1, 0, 0, 0],
+                [0, 1, 0, 0],
+                [0, 0, 1 / np.sqrt(2), -1 / np.sqrt(2)],
+                [0, 0, 1 / np.sqrt(2), 1 / np.sqrt(2)],
+            ]
+        ),
+    ),
+    (qml.CRY, [np.pi], np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, -1], [0, 0, 1, 0]])),
+    (
+        qml.CRY,
+        [np.array([np.pi / 2, np.pi])],
+        np.array(
+            [
+                [
+                    [1, 0, 0, 0],
+                    [0, 1, 0, 0],
+                    [0, 0, 1 / np.sqrt(2), -1 / np.sqrt(2)],
+                    [0, 0, 1 / np.sqrt(2), 1 / np.sqrt(2)],
+                ],
+                [[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, -1], [0, 0, 1, 0]],
+            ]
+        ),
+    ),
+    (qml.CRZ, [0], np.identity(4)),
+    (
+        qml.CRZ,
+        [np.pi / 2],
+        np.array(
+            [
+                [1, 0, 0, 0],
+                [0, 1, 0, 0],
+                [0, 0, np.exp(-1j * np.pi / 4), 0],
+                [0, 0, 0, np.exp(1j * np.pi / 4)],
+            ]
+        ),
+    ),
+    (qml.CRZ, [np.pi], np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, -1j, 0], [0, 0, 0, 1j]])),
+    (
+        qml.CRZ,
+        [np.array([np.pi / 2, np.pi])],
+        np.array(
+            [
+                [
+                    [1, 0, 0, 0],
+                    [0, 1, 0, 0],
+                    [0, 0, np.exp(-1j * np.pi / 4), 0],
+                    [0, 0, 0, np.exp(1j * np.pi / 4)],
+                ],
+                [[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, -1j, 0], [0, 0, 0, 1j]],
+            ]
+        ),
+    ),
+    (qml.CRot, [0, 0, 0], np.identity(4)),
+    (
+        qml.CRot,
+        [np.pi, np.pi, np.pi],
+        np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, -1], [0, 0, 1, 0]]),
+    ),
+    (qml.CRot, [0.432, -0.152, 0.9234], _arbitrary_crot(0.432, -0.152, 0.9234)),
+    (qml.CRot, [np.zeros(5), np.zeros(5), np.zeros(5)], np.identity(4)),
+    (
+        qml.CRot,
+        [np.ones(3) * np.pi, np.ones(3) * np.pi, np.ones(3) * np.pi],
+        np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, -1], [0, 0, 1, 0]]),
+    ),
+    (
+        qml.CRot,
+        [np.array([0.432, -0.124]), np.array([-0.152, 2.912]), np.array([0.9234, -9.2])],
+        _arbitrary_crot_broadcasted(
+            np.array([0.432, -0.124]), np.array([-0.152, 2.912]), np.array([0.9234, -9.2])
+        ),
+    ),
+    (qml.ControlledPhaseShift, [0.123], ControlledPhaseShift(0.123)),
+    (
+        qml.ControlledPhaseShift,
+        [np.array([0.2, np.pi / 2, -0.1])],
+        _phase_shift_matrix_broadcasted(np.array([0.2, np.pi / 2, -0.1])),
+    ),
+]
+
+EXPECTED_EIGVALS = [
+    (qml.CRZ, [0], np.ones(4)),
+    (qml.CRZ, [np.pi / 2], np.array([1, 1, np.exp(-1j * np.pi / 4), np.exp(1j * np.pi / 4)])),
+    (qml.CRZ, [np.pi], np.array([1, 1, -1j, 1j])),
+    (
+        qml.CRZ,
+        [np.array([np.pi / 2, np.pi])],
+        [
+            np.array([1, 1, np.exp(-1j * np.pi / 4), np.exp(1j * np.pi / 4)]),
+            np.array([1, 1, -1j, 1j]),
+        ],
+    ),
+    (qml.ControlledPhaseShift, [0.123], np.linalg.eigvals(ControlledPhaseShift(0.123))),
+    (
+        qml.ControlledPhaseShift,
+        [np.array([0.2, np.pi / 2, -0.1])],
+        np.linalg.eigvals(_phase_shift_matrix_broadcasted(np.array([0.2, np.pi / 2, -0.1]))),
+    ),
+]
+
+
+class TestComputations:
+    """Tests computing the matrix and eigenvalues of a controlled operation"""
+
+    @pytest.mark.parametrize("op, params, expected_matrix", EXPECTED_MATRICES)
+    def test_matrix(self, tol, op, params, expected_matrix):
+        """Tests that the correct matrix is returned"""
+
+        assert np.allclose(op.compute_matrix(*params), expected_matrix, atol=tol, rtol=0)
+        assert np.allclose(op(*params, wires=[0, 1]).matrix(), expected_matrix, atol=tol, rtol=0)
+
+    @pytest.mark.tf
+    @pytest.mark.parametrize("op, params, expected_matrix", EXPECTED_MATRICES)
+    def test_matrix_tf(self, tol, op, params, expected_matrix):
+        """Tests that the correct matrix is returned when using TensorFlow"""
+
+        import tensorflow as tf
+
+        variables = [tf.Variable(param) for param in params]
+
+        assert np.allclose(op.compute_matrix(*variables), expected_matrix, atol=tol, rtol=0)
+        assert np.allclose(op(*variables, wires=[0, 1]).matrix(), expected_matrix, atol=tol, rtol=0)
+
+    @pytest.mark.parametrize("op, params, expected_eigvals", EXPECTED_EIGVALS)
+    def test_eigvals(self, tol, op, params, expected_eigvals):
+        """Tests that the correct eigenvalues are returned"""
+
+        assert np.allclose(op.compute_eigvals(*params), expected_eigvals, atol=tol, rtol=0)
+        assert np.allclose(op(*params, wires=[0, 1]).eigvals(), expected_eigvals, atol=tol, rtol=0)
+
+    @pytest.mark.tf
+    @pytest.mark.parametrize("op, params, expected_eigvals", EXPECTED_EIGVALS)
+    def test_eigvals_tf(self, tol, op, params, expected_eigvals):
+        """Tests that the correct eigenvalues are returned when using TensorFlow"""
+
+        import tensorflow as tf
+
+        variables = [tf.Variable(param) for param in params]
+
+        assert np.allclose(op.compute_eigvals(*variables), expected_eigvals, atol=tol, rtol=0)
+        assert np.allclose(
+            op(*variables, wires=[0, 1]).eigvals(), expected_eigvals, atol=tol, rtol=0
+        )
+
+
+def _dot_broadcasted(a, b):
+    return np.einsum("...ij,...jk->...ik", a, b)
+
+
+def _multi_dot_broadcasted(matrices):
+    return functools.reduce(_dot_broadcasted, matrices)
+
+
+class TestDecompositions:
+    @pytest.mark.parametrize("phi", [0.432, np.array([0.1, 2.1])])
+    def test_CRX(self, phi):
+        """Test the decomposition for CRX."""
+
+        ops1 = qml.CRX.compute_decomposition(phi, wires=[0, 1])
+        ops2 = qml.CRX(phi, wires=(0, 1)).decomposition()
+
+        expected_ops = [
+            qml.RZ(np.pi / 2, 1),
+            qml.RY(phi / 2, 1),
+            qml.CNOT(wires=(0, 1)),
+            qml.RY(-phi / 2, 1),
+            qml.CNOT(wires=(0, 1)),
+            qml.RZ(-np.pi / 2, 1),
+        ]
+
+        assert ops1 == expected_ops
+        assert ops2 == expected_ops
+
+    @pytest.mark.parametrize("phi", [0.432, np.array([2.1, 0.2])])
+    def test_CRY(self, phi):
+        """Test the decomposition for CRY."""
+
+        ops1 = qml.CRY.compute_decomposition(phi, wires=[0, 1])
+        ops2 = qml.CRY(phi, wires=(0, 1)).decomposition()
+
+        expected_ops = [
+            qml.RY(phi / 2, 1),
+            qml.CNOT(wires=(0, 1)),
+            qml.RY(-phi / 2, 1),
+            qml.CNOT(wires=(0, 1)),
+        ]
+
+        assert ops1 == expected_ops
+        assert ops2 == expected_ops
+
+    @pytest.mark.parametrize("phi", [0.321, np.array([0.6, 2.1])])
+    def test_CRZ(self, phi):
+        """Test the decomposition for CRZ."""
+
+        ops1 = qml.CRZ.compute_decomposition(phi, wires=[1, 0])
+        ops2 = qml.CRZ(phi, wires=(1, 0)).decomposition()
+
+        expected_ops = [
+            qml.PhaseShift(phi / 2, wires=0),
+            qml.CNOT(wires=[1, 0]),
+            qml.PhaseShift(-phi / 2, wires=0),
+            qml.CNOT(wires=[1, 0]),
+        ]
+
+        assert ops1 == expected_ops
+        assert ops2 == expected_ops
+
+    @pytest.mark.parametrize("phi, theta, omega", [[0.5, 0.6, 0.7], [0.1, -0.4, 0.7], [-10, 5, -1]])
+    def test_CRot(self, tol, phi, theta, omega):
+        """Tests that the decomposition of the CRot gate is correct"""
+
+        op = qml.CRot(phi, theta, omega, wires=[0, 1])
+        res = op.decomposition()
+
+        mats = []
+        for i in reversed(res):
+            if len(i.wires) == 1:
+                mats.append(np.kron(np.eye(2), i.matrix()))
+            else:
+                mats.append(i.matrix())
+
+        decomposed_matrix = np.linalg.multi_dot(mats)
+        assert np.allclose(decomposed_matrix, op.matrix(), atol=tol, rtol=0)
+
+    @pytest.mark.parametrize(
+        "phi, theta, omega",
+        [
+            [np.array([0.1, 0.2]), np.array([-0.4, 2.19]), np.array([0.7, -0.7])],
+            [np.array([0.1, 0.2, 0.9]), -0.4, np.array([0.7, 0.0, -0.7])],
+        ],
+    )
+    def test_CRot_broadcasted(self, tol, phi, theta, omega):
+        """Tests that the decomposition of the broadcasted CRot gate is correct"""
+
+        op = qml.CRot(phi, theta, omega, wires=[0, 1])
+        res = op.decomposition()
+
+        mats = []
+        for i in reversed(res):
+            mat = i.matrix()
+            if len(i.wires) == 1:
+                I = np.eye(2)[np.newaxis] if qml.math.ndim(mat) == 3 else np.eye(2)
+                mats.append(np.kron(I, mat))
+            else:
+                mats.append(mat)
+
+        decomposed_matrix = _multi_dot_broadcasted(mats)
+        assert np.allclose(decomposed_matrix, op.matrix(), atol=tol, rtol=0)
+
+    @pytest.mark.parametrize("phi", [-0.1, 0.2, 0.5])
+    def test_controlled_phase_shift(self, phi):
+        """Tests that the ControlledPhaseShift calculates the correct decomposition"""
+
+        op = qml.ControlledPhaseShift(phi, wires=[0, 2])
+        decomp = op.decomposition()
+
+        mats = []
+        for i in reversed(decomp):
+            if i.wires.tolist() == [0]:
+                mats.append(np.kron(i.matrix(), np.eye(4)))
+            elif i.wires.tolist() == [1]:
+                mats.append(np.kron(np.eye(2), np.kron(i.matrix(), np.eye(2))))
+            elif i.wires.tolist() == [2]:
+                mats.append(np.kron(np.eye(4), i.matrix()))
+            elif isinstance(i, qml.CNOT) and i.wires.tolist() == [0, 1]:
+                mats.append(np.kron(i.matrix(), np.eye(2)))
+            elif isinstance(i, qml.CNOT) and i.wires.tolist() == [0, 2]:
+                mats.append(
+                    np.array(
+                        [
+                            [1, 0, 0, 0, 0, 0, 0, 0],
+                            [0, 1, 0, 0, 0, 0, 0, 0],
+                            [0, 0, 1, 0, 0, 0, 0, 0],
+                            [0, 0, 0, 1, 0, 0, 0, 0],
+                            [0, 0, 0, 0, 0, 1, 0, 0],
+                            [0, 0, 0, 0, 1, 0, 0, 0],
+                            [0, 0, 0, 0, 0, 0, 0, 1],
+                            [0, 0, 0, 0, 0, 0, 1, 0],
+                        ]
+                    )
+                )
+
+        decomposed_matrix = np.linalg.multi_dot(mats)
+        lam = np.exp(1j * phi)
+        exp = np.array(
+            [
+                [1, 0, 0, 0, 0, 0, 0, 0],
+                [0, 1, 0, 0, 0, 0, 0, 0],
+                [0, 0, 1, 0, 0, 0, 0, 0],
+                [0, 0, 0, 1, 0, 0, 0, 0],
+                [0, 0, 0, 0, 1, 0, 0, 0],
+                [0, 0, 0, 0, 0, lam, 0, 0],
+                [0, 0, 0, 0, 0, 0, 1, 0],
+                [0, 0, 0, 0, 0, 0, 0, lam],
+            ]
+        )
+
+        assert np.allclose(decomposed_matrix, exp)
+
+    def test_controlled_phase_shift_broadcasted(self):
+        """Tests that the ControlledPhaseShift calculates the correct decomposition"""
+
+        phi = np.array([-0.2, 4.2, 1.8])
+        op = qml.ControlledPhaseShift(phi, wires=[0, 2])
+        decomp = op.decomposition()
+
+        mats = []
+        for i in reversed(decomp):
+            mat = i.matrix()
+            eye = np.eye(2)[np.newaxis] if np.ndim(mat) == 3 else np.eye(2)
+            if i.wires.tolist() == [0]:
+                mats.append(np.kron(mat, np.kron(eye, eye)))
+            elif i.wires.tolist() == [1]:
+                mats.append(np.kron(eye, np.kron(mat, eye)))
+            elif i.wires.tolist() == [2]:
+                mats.append(np.kron(np.kron(eye, eye), mat))
+            elif isinstance(i, qml.CNOT) and i.wires.tolist() == [0, 1]:
+                mats.append(np.kron(mat, eye))
+            elif isinstance(i, qml.CNOT) and i.wires.tolist() == [0, 2]:
+                mats.append(
+                    np.array(
+                        [
+                            [1, 0, 0, 0, 0, 0, 0, 0],
+                            [0, 1, 0, 0, 0, 0, 0, 0],
+                            [0, 0, 1, 0, 0, 0, 0, 0],
+                            [0, 0, 0, 1, 0, 0, 0, 0],
+                            [0, 0, 0, 0, 0, 1, 0, 0],
+                            [0, 0, 0, 0, 1, 0, 0, 0],
+                            [0, 0, 0, 0, 0, 0, 0, 1],
+                            [0, 0, 0, 0, 0, 0, 1, 0],
+                        ]
+                    )
+                )
+
+        decomposed_matrix = _multi_dot_broadcasted(mats)
+        lam = np.exp(1j * phi)
+        exp = np.array([np.diag([1, 1, 1, 1, 1, el, 1, el]) for el in lam])
+
+        assert np.allclose(decomposed_matrix, exp)
+
+    @pytest.mark.parametrize("ops, expected_ops", NON_PARAMETRIC_OPS_DECOMPOSITIONS)
+    def test_non_parametric_op_decompositions(self, ops, expected_ops, tol):
+        """Tests that decompositions of non-parametrized operations are correct"""
+
+        op = ops(wires=[0, 1])
+        decomps = op.decomposition()
+        decomposed_matrix = qml.matrix(op.decomposition)()
+
+        for gate, expected in zip(decomps, expected_ops):
+            assert qml.equal(gate, expected, atol=tol, rtol=0)
+
+        assert np.allclose(decomposed_matrix, op.matrix(), atol=tol, rtol=0)
+
+
+def test_simplify_crot():
+    """Simplify CRot operations with different parameters."""
+
+    crot_x = qml.CRot(np.pi / 2, 0.1, -np.pi / 2, wires=[0, 1])
+    simplify_crot_x = crot_x.simplify()
+
+    assert simplify_crot_x.name == "CRX"
+    assert simplify_crot_x.data == (0.1,)
+    assert np.allclose(simplify_crot_x.matrix(), crot_x.matrix())
+
+    crot_y = qml.CRot(0, 0.1, 0, wires=[0, 1])
+    simplify_crot_y = crot_y.simplify()
+
+    assert simplify_crot_y.name == "CRY"
+    assert simplify_crot_y.data == (0.1,)
+    assert np.allclose(simplify_crot_y.matrix(), crot_y.matrix())
+
+    crot_z = qml.CRot(0.1, 0, 0.2, wires=[0, 1])
+    simplify_crot_z = crot_z.simplify()
+
+    assert simplify_crot_z.name == "CRZ"
+    assert np.allclose(simplify_crot_z.data, [0.3])
+    assert np.allclose(simplify_crot_z.matrix(), crot_z.matrix())
+
+    crot = qml.CRot(0.1, 0.2, 0.3, wires=[0, 1])
+    not_simplified_crot = crot.simplify()
+
+    assert not_simplified_crot.name == "CRot"
+    assert np.allclose(not_simplified_crot.matrix(), crot.matrix())
+
+
+controlled_data = [
+    (qml.RX(1.234, wires=0), qml.CRX(1.234, wires=("a", 0))),
+    (qml.RY(1.234, wires=0), qml.CRY(1.234, wires=("a", 0))),
+    (qml.RZ(1.234, wires=0), qml.CRZ(1.234, wires=("a", 0))),
+    (qml.PhaseShift(1.234, wires=0), qml.ControlledPhaseShift(1.234, wires=("a", 0))),
+    (qml.Rot(1.2, 2.3, 3.4, wires=0), qml.CRot(1.2, 2.3, 3.4, wires=("a", 0))),
+]
+
+
+@pytest.mark.parametrize("base, cbase", controlled_data)
+def test_controlled_method(base, cbase):
+    """Tests the _controlled method for parametric ops."""
+    # pylint: disable=protected-access
+    assert qml.equal(base._controlled("a"), cbase)
